pyproject.toml
[tool.poetry]
name = "lowclouds"
version = "0.1.2"
description = "The lowclouds is a shortcut library for several cloud service libraries."
readme = "README.md"
authors = ["rhoboro <rhoboro@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
google-cloud-bigquery = {version = "^1.27.2", optional = true}
google-cloud-bigquery-storage = {version = "^1.0.0", optional = true}
grpcio = {version = "^1.31.0", optional = true}
boto3 = {version = "^1.14.55", optional = true}
google-cloud-storage = {version = "^1.31.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.782"
tox = "^3.20.0"
pytest-mock = "^3.3.1"

[tool.poetry.extras]
bigquery=["google-cloud-bigquery", "google-cloud-bigquery-storage", "grpcio"]
gcs=["google-cloud-storage"]
s3=["boto3"]

[tool.black]
line-length = 100

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

